Make generate_remote_keypair more generic for potential other remote-wallets
